Remove card to just examples
