+++ packages/@react-spectrum/s2/stories/CoachMark.stories.tsx
@@ -12,6 +12,7 @@
 import {
   ActionButton,
   ActionMenu,
+  ActionMenuContext,
   Button,
   ButtonGroup,
   CardPreview,
@@ -19,17 +20,25 @@ import {
   UNSTABLE_CoachMark as CoachMark,
   UNSTABLE_CoachMarkTrigger as CoachMarkTrigger,
   Content,
+  ContentContext,
+  DividerContext,
   Footer,
+  FooterContext,
   Image,
+  ImageContext,
+  ImageCoordinator,
   Keyboard,
+  KeyboardContext,
   MenuItem,
   Slider,
-  Text
+  Text,
+  TextContext
 } from '../src';
 import {card} from '../src/Card';
+import {DEFAULT_SLOT, Provider} from 'react-aria-components';
 import Filter from '../s2wf-icons/S2_Icon_Filter_20_N.svg';
 import type {Meta, StoryObj} from '@storybook/react';
-import {style} from '../style' with {type: 'macro'};
+import {space, style} from '../style' with {type: 'macro'};
 import {useState} from 'react';
 
 const meta: Meta<typeof CoachMark> = {
@@ -56,7 +65,7 @@ export const CoachMarkExample: Story = {
       <CoachMarkTrigger defaultOpen>
         <Checkbox>Sync with CC</Checkbox>
         <CoachMark placement="right top" {...args}>
-          <div className={card({size: 'M', density: 'regular'})}>
+          <CoachMarkCard>
             <CardPreview>
               <Image src={new URL('assets/preview.png', import.meta.url).toString()} />
             </CardPreview>
@@ -76,7 +85,7 @@ export const CoachMarkExample: Story = {
                 <Button variant="primary">Next</Button>
               </ButtonGroup>
             </Footer>
-          </div>
+          </CoachMarkCard>
         </CoachMark>
       </CoachMarkTrigger>
       <Button>After</Button>
@@ -98,7 +107,7 @@ function ControlledCoachMark(args) {
       <CoachMarkTrigger isOpen={isOpen} onOpenChange={setIsOpen}>
         <Checkbox>Sync with CC</Checkbox>
         <CoachMark placement="right top" {...args}>
-          <div className={card({size: 'M', density: 'regular'})}>
+          <CoachMarkCard>
             <CardPreview>
               <Image src={new URL('assets/preview.png', import.meta.url).toString()} />
             </CardPreview>
@@ -111,7 +120,7 @@ function ControlledCoachMark(args) {
               <Keyboard>Command + B</Keyboard>
               <Text slot="description">This is the description</Text>
             </Content>
-          </div>
+          </CoachMarkCard>
         </CoachMark>
       </CoachMarkTrigger>
       <Button onPress={() => setIsOpen(false)}>Close</Button>
@@ -137,7 +146,7 @@ export const CoachMarkSlider: Story = {
       <CoachMarkTrigger defaultOpen>
         <Slider label="Horizontal position" labelPosition="top" />
         <CoachMark placement="right top" {...args}>
-          <div className={card({size: 'M', density: 'regular'})}>
+          <CoachMarkCard>
             <CardPreview>
               <Image src={new URL('assets/preview.png', import.meta.url).toString()} />
             </CardPreview>
@@ -157,7 +166,7 @@ export const CoachMarkSlider: Story = {
                 <Button variant="primary">Next</Button>
               </ButtonGroup>
             </Footer>
-          </div>
+          </CoachMarkCard>
         </CoachMark>
       </CoachMarkTrigger>
       <Button>After</Button>
@@ -179,7 +188,7 @@ export const CoachMarkButton: Story = {
           <Filter />
         </ActionButton>
         <CoachMark placement="right top" {...args}>
-          <div className={card({size: 'M', density: 'regular'})}>
+          <CoachMarkCard>
             <CardPreview>
               <Image src={new URL('assets/preview.png', import.meta.url).toString()} />
             </CardPreview>
@@ -199,7 +208,7 @@ export const CoachMarkButton: Story = {
                 <Button variant="primary">Next</Button>
               </ButtonGroup>
             </Footer>
-          </div>
+          </CoachMarkCard>
         </CoachMark>
       </CoachMarkTrigger>
       <Button>After</Button>
@@ -211,3 +220,158 @@ export const CoachMarkButton: Story = {
     }
   }
 };
+
+function CoachMarkCard(props) {
+  let {size = 'M'} = props;
+  return (
+    <div className={card({size, density: 'regular'})}>
+      <Provider
+        values={[
+          [ImageContext, {alt: '', styles: image}],
+          [TextContext, {
+            slots: {
+              [DEFAULT_SLOT]: {},
+              title: {styles: title({size})},
+              description: {styles: description({size})},
+              steps: {styles: steps}
+            }
+          }],
+          [KeyboardContext, {styles: keyboard}],
+          [ContentContext, {styles: content({size})}],
+          [DividerContext, {size: 'S'}],
+          [FooterContext, {styles: footer}],
+          [ActionMenuContext, {
+            isQuiet: true,
+            size: actionButtonSize[size],
+            // @ts-ignore
+            'data-slot': 'menu',
+            styles: actionMenu
+          }]
+        ]}>
+        <ImageCoordinator>
+          {props.children}
+        </ImageCoordinator>
+      </Provider>
+    </div>
+  );
+}
+
+
+const image = style({
+  width: 'full',
+  aspectRatio: '[3/2]',
+  objectFit: 'cover',
+  userSelect: 'none',
+  pointerEvents: 'none'
+});
+
+let title = style({
+  font: 'title',
+  fontSize: {
+    size: {
+      XS: 'title-xs',
+      S: 'title-xs',
+      M: 'title-sm',
+      L: 'title',
+      XL: 'title-lg'
+    }
+  },
+  lineClamp: 3,
+  gridArea: 'title'
+});
+
+let description = style({
+  font: 'body',
+  fontSize: {
+    size: {
+      XS: 'body-2xs',
+      S: 'body-2xs',
+      M: 'body-xs',
+      L: 'body-sm',
+      XL: 'body'
+    }
+  },
+  lineClamp: 3,
+  gridArea: 'description'
+});
+
+let keyboard = style({
+  gridArea: 'keyboard',
+  font: 'ui',
+  fontWeight: 'light',
+  color: 'gray-600',
+  background: 'gray-25',
+  unicodeBidi: 'plaintext'
+});
+
+let steps = style({
+  font: 'detail',
+  fontSize: 'detail-sm',
+  alignSelf: 'center'
+});
+
+let content = style({
+  display: 'grid',
+  // By default, all elements are displayed in a stack.
+  // If an action menu is present, place it next to the title.
+  gridTemplateColumns: {
+    default: ['1fr'],
+    ':has([data-slot=menu])': ['minmax(0, 1fr)', 'auto']
+  },
+  gridTemplateAreas: {
+    default: [
+      'title keyboard',
+      'description keyboard'
+    ],
+    ':has([data-slot=menu])': [
+      'title menu',
+      'keyboard keyboard',
+      'description description'
+    ]
+  },
+  columnGap: 4,
+  flexGrow: 1,
+  alignItems: 'baseline',
+  alignContent: 'space-between',
+  rowGap: {
+    size: {
+      XS: 4,
+      S: 4,
+      M: space(6),
+      L: space(6),
+      XL: 8
+    }
+  },
+  paddingTop: {
+    default: '--card-spacing',
+    ':first-child': 0
+  },
+  paddingBottom: {
+    default: '[calc(var(--card-spacing) * 1.5 / 2)]',
+    ':last-child': 0
+  }
+});
+
+let actionMenu = style({
+  gridArea: 'menu',
+  // Don't cause the row to expand, preserve gap between title and description text.
+  // Would use -100% here but it doesn't work in Firefox.
+  marginY: '[calc(-1 * self(height))]'
+});
+
+let footer = style({
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'end',
+  justifyContent: 'space-between',
+  gap: 8,
+  paddingTop: '[calc(var(--card-spacing) * 1.5 / 2)]'
+});
+
+const actionButtonSize = {
+  XS: 'XS',
+  S: 'XS',
+  M: 'S',
+  L: 'M',
+  XL: 'L'
+} as const;
